% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optseq.R
\name{optseq}
\alias{optseq}
\title{Optimal Difference Sequence}
\usage{
optseq(r)
}
\arguments{
\item{r}{the order of the generated difference sequence.}
}
\value{
The generated optimal difference sequence.
}
\references{
Hall, P., Kay, J. W. and Titterington, D. M. (1990). Asymptotically optimal difference-based estimation of variance in nonparametric regression, Biometrika 77: 521 - 528.
}
\description{
Generate an optimal difference sequence with order r(<=10).
}
\examples{
r<-2
optseq(r)
}

