% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ordseq.R
\name{ordseq}
\alias{ordseq}
\title{Ordinary Difference Sequence}
\usage{
ordseq(r)
}
\arguments{
\item{r}{the order of the generated difference sequence.}
}
\value{
The generated ordinary difference sequence.
}
\references{
Hall, P., Kay, J. W. and Titterington, D. M. (1990). Asymptotically optimal difference-based estimation of variance in nonparametric regression, Biometrika 77: 521 - 528.

Dette, H., Munk, A. and Wagner, T. (1998). Estimating the variance in nonparametric regression - what is a reasonable choice?, Journal of the Royal Statistical Society, Series B 60: 751 - 764.
}
\description{
Generate an ordinary difference sequence with order r.
}
\examples{
r<-2
ordseq(r)
}

